401 not authorised

+++ controllers/contactController.js
@@ -3,17 +3,17 @@ const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contactModel");
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 
 const getContacts = asyncHandler(async (req,res) => {
     //MODELS
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({user_id: req.user.id});;
     res.status(200).json(contacts);
 });
 
 //@desc Create new contact
 //@route POST /api/contacts
-//@access public
+//@access private
 
 const createContact = asyncHandler(async(req,res) => {
     console.log("The request body is:", req.body);
@@ -24,14 +24,14 @@ const createContact = asyncHandler(async(req,res) => {
     }
     //MODELS - kontakt koji stvaram iz posta
     const contact = await Contact.create({
-        name, email, phone,
+        name, email, phone, user_id: req.user.id
     });
     res.status(201).json(contact);
 });
 
 //@desc Get induvidual contact/:id
 //@route GET /api/contacts
-//@access public
+//@access private
 
 const getInduvidualContact = asyncHandler(async(req,res) => {
 
@@ -47,7 +47,7 @@ const getInduvidualContact = asyncHandler(async(req,res) => {
 
 //@desc Delete  contact
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 
 const deleteContact = asyncHandler(async(req,res) => {
     const contact = await Contact.findById(req.params.id);
@@ -56,6 +56,10 @@ const deleteContact = asyncHandler(async(req,res) => {
         throw new Error ("Contact not found");
 
     }
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error ("Wrong user no permission");
+    } 
     const removedContact = await Contact.findByIdAndRemove(
         req.params.id
     )
@@ -64,7 +68,7 @@ const deleteContact = asyncHandler(async(req,res) => {
 
 //@desc Update  contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 
 const updateContact = asyncHandler(async(req,res) => {
     const contact = await Contact.findById(req.params.id);
@@ -73,7 +77,10 @@ const updateContact = asyncHandler(async(req,res) => {
         throw new Error ("Contact not found");
 
     }
-    const updatedContact = await Contact.findByIdAndUpdate(
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error ("Wrong user no permission");
+    }    const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id, 
         req.body,
         {new:true}
